backend/requirements/base.txt
# Core Django and DRF
Django>=5.1.0,<5.2.0
djangorestframework>=3.16.0,<3.17.0
djangorestframework-simplejwt>=5.0.0,<5.1.0
django-cors-headers>=4.0.0,<4.1.0

# Caching
redis>=5.0.0,<5.3.0
django-redis>=5.4.0,<5.5.0
hiredis>=2.0.0,<3.0.0

# API documentation
drf-yasg>=1.21.0,<1.22.0

# Environment variables
python-dotenv>=1.0.0,<1.2.0

# Utilities
asgiref>=3.8.0,<3.9.0
pytz>=2025.1,<2026.0
sqlparse>=0.4.0,<0.6.0
tzdata>=2025.1,<2026.0
PyYAML>=6.0,<6.1
uritemplate>=4.1.0,<4.2.0
inflection>=0.5.0,<0.6.0
packaging>=24.0,<25.0